+++ packages/config/__tests__/ProfileInfo.TeamConfig.test.ts
@@ -369,6 +369,39 @@ describe("TeamConfig ProfileInfo tests", () => {
             expect(actualDefaultProfiles).toEqual(expectedDefaultProfiles);
             expect(expectedProfileNames.length).toEqual(0);
         });
+
+        it("should return some profiles if a type is specified and exclude homeDir", async () => {
+            const desiredProfType = "zosmf";
+            const expectedName = "LPAR1";
+            const expectedDefaultProfiles = 1;
+            let expectedProfileNames = ["LPAR1", "LPAR2", "LPAR3"];
+            let actualDefaultProfiles = 0;
+
+            const profInfo = createNewProfInfo(teamProjDir);
+            await profInfo.readProfilesFromDisk({homeDir: teamHomeProjDir});
+            const profAttrs = profInfo.getAllProfiles(desiredProfType, {excludeHomeDir: true});
+
+            expect(profAttrs.length).toEqual(expectedProfileNames.length);
+            for (const prof of profAttrs) {
+                if (prof.isDefaultProfile) {
+                    expect(prof.profName).toEqual(expectedName);
+                    actualDefaultProfiles += 1;
+                }
+                expect(expectedProfileNames).toContain(prof.profName);
+                expect(profileTypes).toContain(prof.profType);
+                expect(prof.profLoc.locType).toEqual(ProfLocType.TEAM_CONFIG);
+                expect(prof.profLoc.osLoc).toBeDefined();
+                expect(prof.profLoc.osLoc.length).toEqual(1);
+                expect(prof.profLoc.osLoc[0]).toEqual(path.join(teamProjDir, testAppNm + ".config.json"));
+                expect(prof.profLoc.jsonLoc).toBeDefined();
+                const propertiesJson = jsonfile.readFileSync(path.join(teamProjDir, testAppNm + ".config.json"));
+                expect(lodash.get(propertiesJson, prof.profLoc.jsonLoc)).toBeDefined();
+
+                expectedProfileNames = expectedProfileNames.filter(obj => obj !== prof.profName);
+            }
+            expect(actualDefaultProfiles).toEqual(expectedDefaultProfiles);
+            expect(expectedProfileNames.length).toEqual(0);
+        });
     });
 
     describe("mergeArgsForProfile", () => {