+++ packages/config/src/ProfileInfo.ts
@@ -48,6 +48,7 @@ import {
 } from "../../rest";
 import { IProfInfoUpdateKnownPropOpts, IProfInfoUpdatePropOpts } from "./doc/IProfInfoUpdatePropOpts";
 import { ConfigAutoStore } from "./ConfigAutoStore";
+import { IGetAllProfilesOptions } from "./doc/IProfInfoProps";
 
 /**
  * This class provides functions to retrieve profile-related information.
@@ -318,19 +319,19 @@ export class ProfileInfo {
      *          no profiles of any kind exist), we return an empty array
      *          ie, length is zero.
      */
-    public getAllProfiles(profileType?: string): IProfAttrs[] {
+    public getAllProfiles(profileType?: string, options?: IGetAllProfilesOptions): IProfAttrs[] {
         this.ensureReadFromDisk();
         const profiles: IProfAttrs[] = [];
 
         // Do we have team config profiles?
         if (this.mUsingTeamConfig) {
-            const teamConfigProfs = this.mLoadedConfig.maskedProperties.profiles;
+            const teamConfigProfs = this.mLoadedConfig.layerMerge(true, options?.excludeHomeDir).profiles;
             // Iterate over them
             for (const prof in teamConfigProfs) {
                 // Check if the profile has a type
                 if (teamConfigProfs[prof].type && (profileType == null || teamConfigProfs[prof].type === profileType)) {
                     const jsonLocation: string = "profiles." + prof;
-                    const teamOsLocation: string[] = this.findTeamOsLocation(jsonLocation);
+                    const teamOsLocation: string[] = this.findTeamOsLocation(jsonLocation, options?.excludeHomeDir);
                     const profAttrs: IProfAttrs = {
                         profName: prof,
                         profType: teamConfigProfs[prof].type,
@@ -1212,14 +1213,15 @@ export class ProfileInfo {
 
     /**
      *
-     * @param jsonPath
-     *              The long form JSON path of the profile we are searching for.
+     * @param jsonPath The long form JSON path of the profile we are searching for.
+     * @param excludeHomeDir The long form JSON path of the profile we are searching for.
      * @returns A string array containing the location of all files containing the specified team profile
      */
-    private findTeamOsLocation(jsonPath: string): string[] {
+    private findTeamOsLocation(jsonPath: string, excludeHomeDir?: boolean): string[] {
         const files: string[] = [];
         const layers = this.mLoadedConfig.layers;
         for (const layer of layers) {
+            if (excludeHomeDir && layer.global) continue;
             if (lodash.get(layer.properties, jsonPath) !== undefined) {
                 files.push(layer.path);
             }